Split `extract` into a module
`extract.rs` had reached 1,590 lines holding four separable concerns, and
the work queued behind it only adds more. It becomes a directory:

  mod.rs       the extractor itself and what a layer does to it
  pipeline.rs  blob to bytes: chunks, the buffer pool, both inflaters
  entry.rs     bytes to entries: the tar walk, whiteouts, links
  file.rs      entries to disk: the parts of `unpack_in` we replace
  tests.rs     the layer level tests, which belong to none of them

No behaviour change: code moved, visibility narrowed to `pub(super)`
where the split exposed it, and nothing else. `Compression` and
`compression_of` stay public because the indexer reads them.

The extracted tree is identical and the timings sit on top of each other
(3.084 s against 3.044 s, min of three interleaved rounds).
